requirements.txt
bleak>=0.22.2,<0.23.0
asyncclick>=8.1.7.2,<9.0.0.0
sqlalchemy>=2.0.36,<3.0.0
fastapi>=0.111.1,<0.112.0
uvicorn>=0.30.1,<0.31.0
python-multipart>=0.0.9,<0.0.10
python-dotenv>=1.0.0,<2.0.0